chore(shared-ini-file-loader): add parseKnownFiles from util-credentials
