+++ packages/shared-ini-file-loader/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./getHomeDir";
 export * from "./loadSharedConfigFiles";
+export * from "./parseKnownFiles";
 export * from "./types";
