+++ packages/util-credentials/src/parse-known-profiles.ts
@@ -1,33 +1,14 @@
-import { loadSharedConfigFiles, SharedConfigInit } from "@aws-sdk/shared-ini-file-loader";
-import { ParsedIniData, SharedConfigFiles } from "@aws-sdk/types";
-
-export interface SourceProfileInit extends SharedConfigInit {
-  /**
-   * The configuration profile to use.
-   */
-  profile?: string;
-
-  /**
-   * A promise that will be resolved with loaded and parsed credentials files.
-   * Used to avoid loading shared config files multiple times.
-   *
-   * @internal
-   */
-  loadedConfig?: Promise<SharedConfigFiles>;
-}
+import {
+  parseKnownFiles as __parseKnownFiles,
+  SourceProfileInit as __SourceProfileInit,
+} from "@aws-sdk/shared-ini-file-loader";
 
 /**
- * Load profiles from credentials and config INI files and normalize them into a
- * single profile list.
- *
- * @internal
+ * @deprecated Use SourceProfileInit from "@aws-sdk/shared-ini-file-loader" instead.
  */
-export const parseKnownFiles = async (init: SourceProfileInit): Promise<ParsedIniData> => {
-  const { loadedConfig = loadSharedConfigFiles(init) } = init;
+export interface SourceProfileInit extends __SourceProfileInit {}
 
-  const parsedFiles = await loadedConfig;
-  return {
-    ...parsedFiles.configFile,
-    ...parsedFiles.credentialsFile,
-  };
-};
+/**
+ * @deprecated Use parseKnownFiles from "@aws-sdk/shared-ini-file-loader" instead.
+ */
+export const parseKnownFiles = __parseKnownFiles;
